// frontend/src/types/roles.ts
export enum UserRoleEnum {
    SUPER_ADMIN = "Super Admin",
    MANAGER = "Manager",
    STAFF = "Staff",
    CASHIER = "Cashier",
    PROCUREMENT = "Procurement",
    ACCOUNTANT = "Accountant",
    PHARMACIST = "Pharmacist",
    WAREHOUSE_STAFF = "Warehouse Staff",
};

// export enum UserRoleEnum {
//     SUPER_ADMIN = "SUPER_ADMIN",
//     MANAGER = "MANAGER",
//     STAFF = "STAFF",
//     CASHIER = "CASHIER",
//     PROCUREMENT = "PROCUREMENT",
//     ACCOUNTANT = "ACCOUNTANT",
//     PHARMACIST = "PHARMACIST",
//     WAREHOUSE_STAFF = "WAREHOUSE_STAFF",
// };